+++ Ejercicios Parte 1/Ejercicio15.py
@@ -1,6 +1,6 @@
-n1 = int(input("Escribe el primer número: "))
-n2 = int(input("Escribe el segundo número: "))
-n3 = int(input("Escribe el tercer número: "))
+n1 = float(input("Escribe el primer número: "))
+n2 = float(input("Escribe el segundo número: "))
+n3 = float(input("Escribe el tercer número: "))
 
 if n1 == n2 == n3:
     print("Los tres números son iguales.")
